MBX-2530 Move Monitoring model
